stage8

+++ src/handlers/route-handlers.ts
@@ -1,6 +1,8 @@
 import os from 'node:os';
 import readline from 'node:readline';
 import path from 'node:path';
+import bcrypt from 'bcrypt';
+import jwt from 'jsonwebtoken';
 import { Readable } from 'node:stream';
 import type { Request, Response } from 'express';
 import { generateId } from '../utils/id-generator.js';
@@ -11,6 +13,15 @@ import { promises as fs } from 'node:fs';
 import { parsePagination } from '../utils/parse-pagination.js';
 import pool from '../db/pg-database.js';
 import type { Comment } from '../types/comment.js';
+import type { JwtConfig, JwtPayload, User } from '../types/user.js';
+import { hashRefreshToken } from '../utils/hash-refresh-token.js';
+import { readCookie } from '../utils/read-cookie.js';
+import { isAuthorized } from '../utils/is-authorized.js';
+
+const TOKEN_SECRET = process.env['TOKEN_SECRET'];
+const ACCESS_TTL = '15m';
+const REFRESH_TTL_DAYS = 7;
+
 
 export async function getTasks(req: Request, res: Response): Promise<void> {
     const { page, limit } = parsePagination(req.query);
@@ -29,6 +40,7 @@ export async function createTask(req: Request, res: Response): Promise<void> {
         title: body.title,
         description: body.description,
         status: 'todo',
+        userId: req.user!.id,
         createdAt: new Date().toISOString(),
         updatedAt: new Date().toISOString(),
     };
@@ -63,6 +75,11 @@ export async function updateTask(req: Request<{ id: Task['id'] }>, res: Response
         return;
     }
 
+    if(!isAuthorized(req.user!.role, req.user!.id, task.userId)) {
+        res.status(403).json({ error: 'Forbidden' });
+        return;
+    }
+
     const updates = req.body;
     const updatedValues = {
         ...task,
@@ -80,6 +97,11 @@ export async function updateTask(req: Request<{ id: Task['id'] }>, res: Response
 
 export async function deleteTask(req: Request<{ id: Task['id'] }>, res: Response): Promise<void> {
     const id = req.params.id;
+
+    if (!isAuthorized(req.user!.role, req.user!.id, req.task!.userId)) {
+        res.status(403).json({ error: 'Forbidden' });
+        return;
+    }
 
     try {
         await deleteTaskQuery(id);
@@ -155,6 +177,7 @@ export async function importTasks(req: Request, res: Response): Promise<void> {
                 title: taskData.title || '',
                 description: taskData.description || '',
                 status: taskData.status || 'todo',
+                userId: taskData.userId || '',
                 createdAt: taskData.createdAt || new Date().toISOString(),
                 updatedAt: new Date().toISOString(),
             };
@@ -260,7 +283,11 @@ export async function getTaskComments(req: Request<{ id: Task['id'] }>, res: Res
 
 export async function getTaskQuery(id: Task['id']) {
     const result = await pool.query<Task>(
-        'SELECT id, title, description, status, created_at AS "createdAt", updated_at AS "updatedAt", user_id FROM tasks WHERE id = $1',
+        `SELECT id, title, description, status, 
+            user_id as "userId", 
+            created_at AS "createdAt", 
+            updated_at AS "updatedAt" 
+        FROM tasks WHERE id = $1`,
         [id]
     );
 
@@ -268,12 +295,12 @@ export async function getTaskQuery(id: Task['id']) {
 }
 
 async function insertTaskQuery (task: Task): Promise<Task> {
-    const { id, title, description, status, createdAt, updatedAt } = task;
+    const { id, title, description, status, userId, createdAt, updatedAt } = task;
     const result = await pool.query<Task>(`
-        INSERT INTO tasks (id, title, description, status, created_at, updated_at)
-        VALUES ($1, $2, $3, $4, $5, $6)
-        RETURNING id, title, description, status, created_at AS "createdAt", updated_at AS "updatedAt"
-    `, [id, title, description, status, createdAt, updatedAt]);
+        INSERT INTO tasks (id, title, description, status, user_id, created_at, updated_at)
+        VALUES ($1, $2, $3, $4, $5, $6, $7)
+        RETURNING id, title, description, status, user_id AS "userId", created_at AS "createdAt", updated_at AS "updatedAt"
+    `, [id, title, description, status, userId, createdAt, updatedAt]);
     return result.rows[0] as Task;
 }
 
@@ -297,14 +324,14 @@ async function getTasksQuery (limit: number, offset: number): Promise<Task[]> {
 }
 
 async function updateTaskQuery (task: Task): Promise<Task> {
-    const { title, description, status, updatedAt, id } = task;
+    const { title, description, status, id } = task;
 
     const result = await pool.query(`
         UPDATE tasks
-        SET title = $1, description = $2, status = $3, updated_at = $4
-        WHERE id = $5
+        SET title = $1, description = $2, status = $3, updated_at = NOW()
+        WHERE id = $4
         RETURNING id, title, description, status, created_at AS "createdAt", updated_at AS "updatedAt"
-    `, [title, description, status, updatedAt, id]);
+    `, [title, description, status, id]);
 
     const updatedTask = result.rows[0];
 
@@ -316,7 +343,7 @@ async function deleteTaskQuery (id: Task['id']): Promise<void> {
 }
 
 async function exportTasksQuery (): Promise<Task[]> {
-    const result = await pool.query('SELECT id, title, description, status, created_at AS "createdAt", updated_at AS "updatedAt" FROM tasks');
+    const result = await pool.query('SELECT id, title, description, status, user_id AS "userId", created_at AS "createdAt", updated_at AS "updatedAt" FROM tasks');
 
     return result.rows;
 }
@@ -370,4 +397,194 @@ async function getTaskCommentsQuery (taskId: Task['id']): Promise<Comment[]> {
     );
 
     return result.rows;
+}
+
+export async function createUser(req: Request, res: Response): Promise<void> {
+    const body = req.body;
+    const SALT_ROUNDS = 12;
+    const hashedPassword = await bcrypt.hash(body.password, SALT_ROUNDS);
+
+    const user: User = {
+        id: generateId(),
+        username: body.username,
+        email: body.email,
+        role: 'user',
+        passwordHash: hashedPassword,
+        createdAt: new Date().toDateString(),
+    }
+
+    try {
+        await insertUserQuery(user);
+    } catch (err) {
+        res.status(500).json({error: err})
+        return;
+    }
+
+    res.status(201).json('User has been created');
+}
+
+async function insertUserQuery(userData: User): Promise<void> {
+    const {id, username, email, role, passwordHash, createdAt} = userData;
+    await pool.query(`
+        INSERT INTO users (id, username, email, role, password_hash, created_at)
+        VALUES ($1, $2, $3, $4, $5, $6)
+    `, [id, username, email, role, passwordHash, createdAt]);
+}
+
+export async function login(req: Request, res: Response): Promise<void>{
+    const { email, password } = req.body;
+
+    const userData: User = await getUserQuery(email);
+
+    if (!userData) {
+        res.status(400).json("Invalid credentials");
+        return;
+    }
+
+    const userId = userData.id;
+    const isPasswordValid = await bcrypt.compare(password, userData.passwordHash);
+
+    if (!isPasswordValid) {
+        res.status(400).json("Invalid credentials");
+        return;
+    }
+
+    if (!TOKEN_SECRET) {
+        res.status(500).json("Server configuration error");
+        return;
+    }
+
+    const tokenPayload: JwtPayload = { userId: userData.id, userRole: userData.role };
+
+    const tokenConfig: JwtConfig = {
+        access: {
+            expiresIn: ACCESS_TTL,
+        },
+        refresh: {
+            expiresIn: `${REFRESH_TTL_DAYS}d`,
+        }
+    }
+
+    const accessToken = jwt.sign(
+        tokenPayload, 
+        TOKEN_SECRET,
+        tokenConfig.access
+    );
+
+    const refreshToken = jwt.sign(
+        tokenPayload, 
+        TOKEN_SECRET,
+        tokenConfig.refresh
+    );
+    const refreshTokenHash = hashRefreshToken(refreshToken);
+
+    try {
+        await insertRefreshTokenQuery(refreshTokenHash, userId)
+    } catch (err) {
+        res.status(500).json({error: err});
+        return;
+    }
+
+    res.cookie('refreshToken', refreshToken, {
+        httpOnly: true,
+        secure: true,
+        sameSite: 'strict',
+    });
+    res.json({accessToken});
+}
+
+async function getUserQuery(email: User['email']) {
+    const result = await pool.query(`
+        SELECT id, email, role, password_hash AS "passwordHash" FROM users WHERE email = $1
+    `, [email]);
+    return result.rows[0];
+}
+
+async function insertRefreshTokenQuery (hash: string, userId: User['id']): Promise<void> {
+    await pool.query(`
+        INSERT INTO tokens (id, user_id, token_hash, expires_at)
+        VALUES ($1, $2, $3, NOW() + INTERVAL '7 days');
+    `, [generateId(), userId, hash]);
+}
+
+export async function refresh(req: Request, res: Response): Promise<void> {
+    if(!TOKEN_SECRET) {
+        res.status(500).json({ error: 'Server configuration error' });
+        return;
+    }
+    
+    const refreshToken = readCookie(req.headers.cookie, 'refreshToken');
+    if(!refreshToken) {
+        res.status(401).json({ error: 'Unauthorized' });
+        return;
+    }
+
+    try {
+        const decoded = jwt.verify(refreshToken, TOKEN_SECRET);
+
+        if (typeof decoded === 'string' || !('userId' in decoded)) {
+            res.status(401).json({ error: 'Unauthorized' });
+            return;
+        }
+
+        const tokenHash = hashRefreshToken(refreshToken);
+        const tokenRow = await getValidRefreshTokenByHashQuery(tokenHash);
+
+        if(!tokenRow) {
+            res.status(401).json({ error: 'Unauthorized' });
+            return;
+        }
+
+        const payload: JwtPayload = {
+            userId: (decoded as JwtPayload).userId,
+            userRole: (decoded as JwtPayload).userRole,
+        };
+        const accessToken = jwt.sign(payload, TOKEN_SECRET, { expiresIn: ACCESS_TTL});
+
+        res.json({ accessToken });
+    } catch {
+        res.status(401).json({ error: 'Unauthorized' });
+    }
+}
+
+async function getValidRefreshTokenByHashQuery(tokenHash: string) {
+    const result = await pool.query(
+        `SELECT id, user_id AS "userId", expires_at AS "expiresAt"
+         FROM tokens
+         WHERE token_hash = $1
+           AND revoked_at IS NULL
+           AND expires_at > NOW()
+         LIMIT 1`,
+        [tokenHash]
+    );
+    return result.rows[0];
+}
+
+export async function logout (req: Request, res: Response) {
+    const refreshToken = readCookie(req.headers.cookie, 'refreshToken');
+
+    if (refreshToken) {
+        const tokenHash = hashRefreshToken(refreshToken);
+        await revokeRefreshTokenByHashQuery(tokenHash);
+    }
+
+    res.clearCookie('refreshToken', {
+        httpOnly: true,
+        secure: process.env['NODE_ENV'] === 'production',
+        sameSite: 'strict',
+    });
+
+    res.status(204).end();
+}
+
+async function revokeRefreshTokenByHashQuery(tokenHash: string): Promise<void> {
+    await pool.query(
+        `
+        UPDATE tokens
+        SET revoked_at = NOW()
+        WHERE token_hash = $1
+          AND revoked_at IS NULL
+        `,
+        [tokenHash]
+    );
 }
